### 2022-05-13 #4358 @flyaflya
Thank you for agreeing to edit this submission.

Very relevant reviewer as maintainer of greta package which causact relies on for computational inference:
njtierney (especially relevant due to causact's dependency on the package Nick maintains)

Reviewers who like Bayesian methods and have R Preference:
zhiiiyang
nhtran93
thomaspinder
rMassimiliano
rowlandseymour
ChristopherLucas

Previous reviewers for related package (i.e. greta)
lionel68
joethorley

Not on list, but very good reviewer as author of greta package:
goldingn

Thanks again for the help.  Reach out if you need more suggestions.
