### 2022-05-13 #4358 @joethorley
@terrytangyuan - I'd very much like to review this package (I like Bayesian methods and have R preference)
