### 2022-05-13 e98158d
make variable a distribution for improper flat priors
