### 2022-05-24 #4415 @editorialbot
[REVIEW]: Generative DAGs as an Interface Into Probabilistic Programming with the R Package causact
**Submitting author:** @flyaflya (<a href="http://orcid.org/0000-0003-2871-4788">Adam Fleischhacker</a>)
**Repository:** https://github.com/flyaflya/causact/
**Branch with paper.md** (empty if default branch):
**Version:** v0.4.1
