### 2022-05-24 #4415 @ChristopherLucas
In Figure 1, why is there an edge from the estimated standard error?
