### 2022-05-27 flyaflya/causact#43 @zhiiiyang
Miscellaneous requests
Hi, I am the reviewer from JOSS (https://github.com/openjournals/joss-reviews/issues/4415). Here are some suggestions for your package (to be continued). I will check it off once you have completed the task.

- [x] add Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
- [x] Please add at least one unit test for all exported functions.
- [x] There is no vignette available for this package. Are [eightSchools.R](https://github.com/flyaflya/causact/blob/master/eightSchools.R) and [gretaExamples.R](https://github.com/flyaflya/causact/blob/master/gretaExamples.R) example scripts? If so, please create a folder called vignettes and move these two folders there.
