### 2022-05-28 flyaflya/causact#45 @joethorley
R code in paper should be edited so it can be copied and pasted as is
for example

```
R> graph = dag_create() %>%
+    dag_node("Treatment Effect","y",
```

should be

```
graph = dag_create() %>%
   dag_node("Treatment Effect","y",
```

etc

links to https://github.com/openjournals/joss-reviews/issues/4415
