### 2022-05-28 #4415 @joethorley
@flyaflya  - this is a great piece of work. Congratulations. Comments to follow:

R code in paper should be edited so it can be copied and pasted as is see
https://github.com/flyaflya/causact/issues/45
