### 2022-05-28 flyaflya/causact#46 @joethorley
I think the paper should acknowledge Doodlebugs in its comparison to other packages
I think this is relevant even if Doodlebugs is no longer supported in anyway

https://github.com/openjournals/joss-reviews/issues/4415
