### 2022-05-28 flyaflya/causact#47 @joethorley
README requires clear community guidelines
https://github.com/openjournals/joss-reviews/issues/4415
