### 2022-05-28 flyaflya/causact#49 @joethorley
Links to references need adding where available
for example

McElreath, R. (2020a). Rethinking: Statistical rethinking book package.

should include a link to https://github.com/rmcelreath/rethinking

https://github.com/openjournals/joss-reviews/issues/4415
