### 2022-05-28 flyaflya/causact#50 @joethorley
README should have lifecycle badge to indicate status
In particular further down the README it states that

```
> NOTE: Package is under active development. Breaking changes are to be expected.  Feedback and encouragement is appreciated via github issues or Twitter (https://twitter.com/preposterior).
```

however this information should be immediately conveyed to users via a lifecycle badge at the top of the README.

See for example https://lifecycle.r-lib.org/articles/stages.html

https://github.com/openjournals/joss-reviews/issues/4415
