### 2022-05-28 flyaflya/causact#51 @joethorley
`?causact` should bring up description of file
To generate run `usethis::use_package_doc()`

https://github.com/openjournals/joss-reviews/issues/4415
