### 2022-05-29 flyaflya/causact#46 @joethorley
I also think some discussion of dagitty is relevant http://dagitty.net
