+++ paper/paper.md
@@ -53,28 +53,28 @@ Generative DAGs pursue two simultaneous goals.  One goal is to capture the narra
 Capturing the narrative in code uses some core `causact` functions like `dag_create()`, `dag_node()`, `dag_edge()`, and `dag_plate()` with the chaining operator `%>%` used to build a DAG from the individual elements.  `dag_render()` or `dag_greta()` are then used to visualize the DAG or run inference on the DAG, respectively.  The simplicity with which generative DAGs are constructed belies the complexity of models which can be supported.  For example, multi-level or hierarchical models are easily constructed as shown here in code for constructing and visualizing an oft-cited Bayesian example known as eight schools [@JSSv012i03] whose data is included in `causact` (`causact::schoolsDF`).  The example is a study of coaching effects on test scores where students from eight schools were put into coached and uncoached groups.
 
 ```
-R> graph = dag_create() %>%
-+    dag_node("Treatment Effect","y",
-+             rhs = normal(theta, sigma),
-+             data = causact::schoolsDF$y) %>%
-+    dag_node("Std Error of Effect Estimates","sigma",
-+             data = causact::schoolsDF$sigma,
-+             child = "y") %>%
-+    dag_node("Exp. Treatment Effect","theta",
-+             child = "y",
-+             rhs = avgEffect + schoolEffect) %>%
-+    dag_node("Population Treatment Effect","avgEffect",
-+             child = "theta",
-+             rhs = normal(0,30)) %>%
-+    dag_node("School Level Effects","schoolEffect",
-+             rhs = normal(0,30),
-+             child = "theta") %>%
-+    dag_plate("Observation","i",nodeLabels = c("sigma","y","theta")) %>%
-+    dag_plate("School Name","school",
-+              nodeLabels = "schoolEffect",
-+              data = causact::schoolsDF$schoolName,
-+              addDataNode = TRUE)
-R> graph %>% dag_render()
+graph = dag_create() %>%
+    dag_node("Treatment Effect","y",
+             rhs = normal(theta, sigma),
+             data = causact::schoolsDF$y) %>%
+    dag_node("Std Error of Effect Estimates","sigma",
+             data = causact::schoolsDF$sigma,
+             child = "y") %>%
+    dag_node("Exp. Treatment Effect","theta",
+             child = "y",
+             rhs = avgEffect + schoolEffect) %>%
+    dag_node("Population Treatment Effect","avgEffect",
+             child = "theta",
+             rhs = normal(0,30)) %>%
+    dag_node("School Level Effects","schoolEffect",
+             rhs = normal(0,30),
+             child = "theta") %>%
+    dag_plate("Observation","i",nodeLabels = c("sigma","y","theta")) %>%
+    dag_plate("School Name","school",
+              nodeLabels = "schoolEffect",
+              data = causact::schoolsDF$schoolName,
+              addDataNode = TRUE)
+graph %>% dag_render()
 ```
 
 <center>
@@ -84,7 +84,7 @@ R> graph %>% dag_render()
 \autoref{fig:eightShort} replicates \autoref{fig:eightSchools} without math for less intimidating discussions with domain experts about the model using the `shortLabel = TRUE` argument (shown below).  `causact` does not require a complete model specification prior to rendering the DAG, hence, `causact` facilitates qualitative collaboration on the model design between less technical domain experts and the model builder.
 
 ```
-R> graph %>% dag_render(shortLabel = TRUE)
+graph %>% dag_render(shortLabel = TRUE)
 ```
 
 <center>
@@ -96,9 +96,9 @@ All visualizations, including \autoref{fig:eightSchools} and \autoref{fig:eightS
[…]
