### 2022-06-02 c93c38f
fix typo (issue #52)
