### 2022-06-02 9320676
README should have lifecycle badge to indicate status #50.
