+++ R/dag_create.R
@@ -1,6 +1,8 @@
-#' Create a graph object focused on drawing a DAG.
+#' Create a graph object for drawing a DAG.
+#' @description
+#' `r lifecycle::badge('stable')`
 #'
-#' Generates a \code{causact_graph} graph object that is set-up for drawing DAG graphs.
+#' Generates a `causact_graph` graph object that is set-up for drawing DAG graphs.
 #' @examples
 #' # With `dag_create()` we can create an empty graph and
 #' # add in nodes (`dag_node()`), add edges (`dag_edge`), and
@@ -8,6 +10,7 @@
 #' dag_create()
 #' @return a list object of class `causact_graph` consisting of 6 dataframes.  Each data frame is responsible for storing information about nodes, edges, plates, and the relationships among them.
 #' @export
+
 dag_create <- function() {
   # Create an empty node data frame (`ndf`)
   ndf <-
