### 2022-06-02 flyaflya/causact#50 @flyaflya
Add a "maturing" badge to the readme as the package has been stable for over a year without breaking changes.  I also labelled each function with a badge to show that most of the functionality is anticipated to be stable.

Fixed in https://github.com/flyaflya/causact/commit/9320676955a8b74f8f9670f6cc154b6d73ac353f
