+++ paper/joss.bib
@@ -1,20 +1,3 @@
-@Book{Cameron+Trivedi:2013,
-  author = {A. Colin Cameron and Pravin K. Trivedi},
-  title = {Regression Analysis of Count Data},
-  year = {2013},
-  edition = {2nd},
-  publisher = {Cambridge University Press},
-  address = {Cambridge},
-}
-
-@Book{Chambers+Hastie:1992,
-  editor = {John M. Chambers and Trevor J. Hastie},
-  title = {Statistical Models in \proglang{S}},
-  publisher = {Chapman \& Hall},
-  year = {1992},
-  address = {London},
-}
-
 @article{gelfand1990sampling,
   title={Sampling-based approaches to calculating marginal densities},
   author={Gelfand, Alan E and Smith, Adrian FM},
@@ -63,113 +46,11 @@
   publisher={Academic Press}
 }
 
-@article{hoffman2014no,
-  title={The No-U-Turn sampler: adaptively setting path lengths in Hamiltonian Monte Carlo.},
-  author={Hoffman, Matthew D and Gelman, Andrew},
-  journal={J. Mach. Learn. Res.},
-  volume={15},
-  number={1},
-  pages={1593--1623},
-  year={2014}
-}
-
-
 @Book{gelman2013bayesian,
   title={{Bayesian} data analysis},
   author={Gelman, Andrew and Carlin, John B and Stern, Hal S and Dunson, David B and Vehtari, Aki and Rubin, Donald B},
   year={2013},
   publisher={CRC press}
-}
-
-
-@Manual{Jackman:2015,
-  title = {\pkg{pscl}: Classes and Methods for \proglang{R} Developed in the Political Science Computational Laboratory, Stanford University},
-  author = {Simon Jackman},
-  year = {2015},
-  note = {\proglang{R} package version 1.4.9},
-  url = {https://CRAN.R-project.org/package=pscl},
-}
-
-@Article{Mullahy:1986,
-  author = {John Mullahy},
-  title = {Specification and Testing of Some Modified Count Data Models},
-  year = {1986},
-  journal = {Journal of Econometrics},
-  volume = {33},
-  number = {3},
-  pages = {341--365},
-  doi = {10.1016/0304-4076(86)90002-3},
-}
-
-@Book{McCullagh+Nelder:1989,
-  author = {Peter McCullagh and John A. Nelder},
-  title = {Generalized Linear Models},
-  edition = {2nd},
-  year = {1989},
-  publisher = {Chapman \& Hall},
-  address = {London},
-  doi = {10.1007/978-1-4899-3242-6},
-}
-
-@Manual{R,
-  title = {\proglang{R}: {A} Language and Environment for Statistical Computing},
-  author = {{\proglang{R} Core Team}},
-  organization = {\proglang{R} Foundation for Statistical Computing},
-  address = {Vienna, Austria},
-  year = {2017},
-  url = {https://www.R-project.org/},
-}
-
-@Article{Stasinopoulos+Rigby:2007,
-  author = {D. Mikis Stasinopoulos and Robert A. Rigby},
-  title = {Generalized Additive Models for Location Scale and Shape ({GAMLSS}) in \proglang{R}},
-  journal = {Journal of Statistical Software},
-  year = {2007},
-  volume = {23},
-  number = {7},
-  pages = {1--46},
-  doi = {10.18637/jss.v023.i07},
-}
-
-@Book{Venables+Ripley:2002,
-  author = {William N. Venables and Brian D. Ripley},
-  title = {Modern Applied Statistics with \proglang{S}},
-  edition = {4th},
-  year = {2002},
-  pages = {495},
-  publisher = {Springer-Verlag},
-  address = {New York},
-  doi = {10.1007/978-0-387-21706-2},
-}
-
[…]
