### 2022-06-02 3d826c7
fixed capitalization and added hyperlinks for references as available.
