+++ paper/joss.bib
@@ -1,184 +1,170 @@
-@article{gelfand1990sampling,
-  title={Sampling-based approaches to calculating marginal densities},
-  author={Gelfand, Alan E and Smith, Adrian FM},
-  journal={Journal of the American statistical association},
-  volume={85},
-  number={410},
-  pages={398--409},
-  year={1990},
-  publisher={Taylor \& Francis Group}
-}
-
-@article{gilks1996strategies,
-  title={Strategies for improving MCMC},
-  author={Gilks, Walter R and Roberts, Gareth O},
-  journal={Markov chain Monte Carlo in practice},
-  volume={6},
-  pages={89--114},
-  year={1996}
-}
-
-@book{congdon2010applied,
-  title={Applied {Bayesian} hierarchical methods},
-  author={Congdon, Peter D},
-  year={2010},
-  publisher={CRC Press}
-}
-
-@book{neal1993probabilistic,
-  title={Probabilistic inference using Markov chain Monte Carlo methods},
-  author={Neal, Radford M},
-  year={1993},
-  publisher={Department of Computer Science, University of Toronto Toronto, Ontario, Canada}
-}
-
-@book{pfeffer2016practical,
-  title={Practical probabilistic programming},
-  author={Pfeffer, Avi},
-  year={2016},
-  publisher={Manning Publ.}
-}
-
-@article{kruschke2014doing,
-  title={Doing {Bayesian} data analysis: A tutorial with R, JAGS, and Stan},
-  author={Kruschke, John},
-  year={2014},
-  publisher={Academic Press}
-}
-
-@Book{gelman2013bayesian,
-  title={{Bayesian} data analysis},
-  author={Gelman, Andrew and Carlin, John B and Stern, Hal S and Dunson, David B and Vehtari, Aki and Rubin, Donald B},
-  year={2013},
-  publisher={CRC press}
-}
-
-@book{koller2009probabilistic,
-  title={Probabilistic graphical models: principles and techniques},
-  author={Koller, Daphne and Friedman, Nir},
-  year={2009},
-  publisher={MIT press}
-}
-
-
 @book{stan_development_team,
-  author = {{Stan Development Team}},
-  title = {Stan {Modeling} {Language} {User}'s {Guide} and {Reference} {Manual}, {Version} 2.26},
-  url = {http://mc-stan.org/},
-  year = 2021
+	title        = {Stan {Modeling} {Language} {User}'s {Guide} and {Reference} {Manual}, {Version} 2.26},
+	author       = {{Stan Development Team}},
+	year         = 2021,
+	url          = {http://mc-stan.org/}
 }
-
 @article{burkner2017brms,
-  title={brms: An R package for {Bayesian} multilevel models using Stan},
-  author={B{\"u}rkner, Paul-Christian},
-  journal={Journal of statistical software},
-  volume={80},
-  number={1},
-  pages={1--28},
-  year={2017}
+	title        = {brms: An R package for {Bayesian} multilevel models using Stan},
+	author       = {B{\"u}rkner, Paul-Christian},
+	year         = 2017,
+	journal      = {Journal of statistical software},
+	volume       = 80,
+	number       = 1,
+	pages        = {1--28},
+	url          = {https://doi.org/10.18637/jss.v080.i01}
 }
-
-@article{golding2019greta,
-  title={greta: simple and scalable statistical modelling in R},
-  author={Golding, Nick},
-  journal={Journal of Open Source Software},
-  volume={4},
-  number={40},
-  pages={1601},
-  year={2019}
-}
-
-@article{dillon2017tensorflow,
-  title={Tensorflow distributions},
[…]
