### 2022-06-02 52bdbdc
paper acknowledges Doodlebugs and DAGitty packages #46
