+++ paper/joss.bib
@@ -168,3 +168,23 @@
 	isbn         = {978-3-319-24277-4},
 	url          = {https://ggplot2.tidyverse.org}
 }
+@article{lunn2000,
+	title        = {{WinBUGS} - A {B}ayesian modelling framework: Concepts, structure, and extensibility},
+	author       = {Lunn, David J. and Thomas, Andrew and Best, Nicky and Spiegelhalter, David},
+	year         = 2000,
+	journal      = {Statistics and Computing},
+	volume       = 10,
+	pages        = {325-337},
+	url = {https://doi.org/10.1023/A:1008929526011}
+}
+@article{textor2017,
+    author = {Textor, Johannes and van der Zander, Benito and Gilthorpe, Mark S and Liśkiewicz, Maciej and Ellison, George TH},
+    title = "{Robust causal inference using directed acyclic graphs: the R package ‘dagitty’}",
+    journal = {International Journal of Epidemiology},
+    volume = {45},
+    number = {6},
+    pages = {1887-1894},
+    year = {2017},
+    month = {01},
+    url = {https://doi.org/10.1093/ije/dyw341}
+}
