### 2022-06-02 1cc0902
add mistakenly deleted link for eight schools example
