+++ paper/joss.bib
@@ -179,7 +179,7 @@
 }
 @article{textor2017,
     author = {Textor, Johannes and van der Zander, Benito and Gilthorpe, Mark S and Liśkiewicz, Maciej and Ellison, George TH},
-    title = "{Robust causal inference using directed acyclic graphs: the R package ‘dagitty’}",
+    title = "{Robust causal inference using directed acyclic graphs: the {R} package ‘dagitty’}",
     journal = {International Journal of Epidemiology},
     volume = {45},
     number = {6},
@@ -188,3 +188,17 @@
     month = {01},
     url = {https://doi.org/10.1093/ije/dyw341}
 }
+
+@article{JSSv012i03,
+   author = {Sibylle Sturtz and Uwe Ligges and Andrew Gelman},
+   title = {R2WinBUGS: A Package for Running WinBUGS from R},
+   journal = {Journal of Statistical Software, Articles},
+   volume = {12},
+   number = {3},
+   year = {2005},
+   issn = {1548-7660},
+   pages = {1--16},
+   doi = {10.18637/jss.v012.i03},
+   url = {https://www.jstatsoft.org/v012/i03}
+}
+
