+++ .github/contributing.md
@@ -1,0 +1,17 @@
+## New contributor guide
+
+To get an overview of the project, read the [README](README.md) and Chapters 15-24 of [A Business Analyst's Introduction to Business Analytics](http://www.causact.com). 
+
+### Use GitHub for Communicating Issues and Enhancement Requests
+
+#### Create a new issue
+
+If you are having trouble or find a bug in the package or documentation, [search if an issue already exists](https://github.com/flyaflya/causact/issues). If a related issue doesn't exist, please open a new issue. 
+
+#### Solve an issue
+
+Scan through our [existing issues](https://github.com/flyaflya/causact/issues) to find one that interests you. You can narrow down the search using `labels` as filters.If you find an issue to work on, you are welcome to create a pull request, also known as a PR, with a fix.
+
+##### Pull Request
+
+When you're finished with the changes, create a pull request, also known as a PR.  Thanks for helping!!
