+++ .github/contributing.md
@@ -10,8 +10,13 @@ If you are having trouble or find a bug in the package or documentation, [search
 
 #### Solve an issue
 
-Scan through our [existing issues](https://github.com/flyaflya/causact/issues) to find one that interests you. You can narrow down the search using `labels` as filters.If you find an issue to work on, you are welcome to create a pull request, also known as a PR, with a fix.
+Scan through our [existing issues](https://github.com/flyaflya/causact/issues) to find one that interests you. You can narrow down the search using `labels` as filters. If you find an issue to work on, you are welcome to create a pull request, also known as a PR, with a fix.
 
-##### Pull Request
+#### Submitting a Pull Request
+Before submitting a Pull Request please first open an issue to get feedback on the change you intend to submit.
 
-When you're finished with the changes, create a pull request, also known as a PR.  Thanks for helping!!
+1. Most code should include unit tests (and sometimes e2e tests). New features should include new tests in the same PR. And changes to existing behaviour should update the relevant tests.
+
+2. Most PR's should be linked to one or more relevant issues that they implement.
+
+Thanks for helping!!
