### 2022-06-02 cd691fa
update readme.md with contribution guidelines and getting help / reporting issues guidelines.
