+++ README.Rmd
@@ -128,6 +128,17 @@ drawsDF  ### see top of data frame
 drawsDF %>% dagp_plot()
 ```
 
+
+## Getting help
+
+Whether you encounter a clear bug or just have a question, we are thrilled to help you out.  In either case, please file a [GitHub issue](https://github.com/causact/issues).  If reporting a bug, please include a minimal reproducible example.  If encountering issues installing `greta`, please seek help at the [greta discussion forum](https://forum.greta-stats.org/).
+
+## Contributing
+
+We welcome help turning `causact` into the most intuitive and fastest method of converting stakeholder narratives about data-generating processes into actionable insight from posterior distributions.  If you want to help us achieve this vision, we welcome your contributions.  Please note that this project is released with a [Contributor Code of
+Conduct](https://github.com/flyaflya/causact/blob/master/CODE_OF_CONDUCT.md). By participating
+in this project you agree to abide by its terms.
+
 ## Further Usage
 
 For more info, see `A Business Analyst's Introduction to Business Analytics` available at https://www.causact.com.  Two additional examples are shown below.
@@ -359,14 +370,3 @@ drawsDF %>% dagp_plot()
 ```{r mixtureGraphRenderPost, out.width = "100%", echo = FALSE}
 knitr::include_graphics("man/figures/mixturePost.png")
 ```
-
-## Getting help
-
-Whether you encounter a clear bug or just have a question, we are thrilled to help you out.  In either case, please file a [GitHub issue](https://github.com/causact/issues).  If reporting a bug, please include a minimal reproducible example.  If encountering issues installing `greta`, please seek help at the [greta discussion forum](https://forum.greta-stats.org/).
-
-------------------------------------------------------------------------
-
-Please note that this project is released with a [Contributor Code of
-Conduct](https://github.com/flyaflya/causact/blob/master/CODE_OF_CONDUCT.md). By participating
-in this project you agree to abide by its terms.
-
