### 2022-06-02 26f379f
further edits to readme for getting help, submitting issues, and contributing
