### 2022-06-02 #4415 @flyaflya
@zhiiiyang .  Thx for agreeing to review and I am working on your checklist.  Below, I want to address your question about the second author's contribution.

The second author is my PhD advisee who has been working on this project alongside me since 2019.  She authored the first draft of our submitted paper and has contributed to the development and thought process behind the `causact` package by developing a parallel backend using Julia instead of Tensorflow/Python.  Her main code contributions have not been merged as I am not ready to support the Julia backend quite yet (just supporting the `greta` backend for now).    You can see evidence of her extensive coding efforts here: https://github.com/flyaflya/causact/compare/master...Rosenguyen0411:master.
