### 2022-06-02 #4415 @flyaflya
@joethorley - Thanks a million for all the helpful suggestions.  I think I have addressed your concerns and I went ahead and closed all the issues with comments regarding the commits that fix the raised issues.  Please have a look to make sure they are all okay.  I can reopen any issues that were not resolved properly or need further improvement (and my apologies if I should not have closed them ... not sure of the protocol for whether I close them when I think they are complete or when you say they are resolved satisfactorily).
