### 2022-06-03 #4415 @ChristopherLucas
>

It is possible I'm misunderstanding the example or the DAG, and maybe @joethorley or @zhiiiyang can weigh in, but I don't understand why there's an edge from the _estimated_ SE. E.G., the treatment effect is not dependent on my estimates of parameters of the sampling distribution of the estimator.

Again, it's possible I'm just misunderstanding the example, but presumably others might as well.
