### 2022-06-04 #4415 @joethorley
@ChristopherLucas My understanding is that the edge is there because the values of the treatment effect depends on the expected treatment effect and the standard deviation of the variation in the treatment effect.
