### 2022-06-07 77229b2
add vignette vignette("narrative-to-insight-with-causact")
