+++ vignettes/narrative-to-insight-with-causact.Rmd
@@ -1,0 +1,242 @@
+---
+title: "causact: From narrative to computational insight"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{narrative-to-insight-with-causact}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  warning = FALSE, message = FALSE, 
+  comment = "#>"
+)
+```
+
+This vignette illustrates how to use the `causact` package to reduce the friction experienced while going from business issues to computational Bayesian models, inference, and insight.  The package is easy-to-use, intuitive, and provides a visual interface to `greta`, an `R` based probabilistic programming language (PPL).  To get started load the required packages:
+
+```{r setup}
+library(greta)
+library(causact)
+```
+
+## Motivating Business Narrative
+
+We will use the below example fictional narrative along with some simulated data to motivate how to use use `causact` for iterating from narrative to computational insight.  A more detailed version of this vignette can be found in Chapters 14-17 of "A Business Analyst's Introduction to Business Analytics" available at http://www.causact.com.
+
+> BankPass is an auto financing company that is launching a new credit card, the Travel Rewards Card (TRC).  The card is designed for people who love to take adventure vacations like ziplining, kayaking, scuba diving, and the like.  To date, BankPass has been randomly selecting a small number of customers to target with a private offer to sign up for the card.  They would like to assess the probability that an individual will get a TRC card if exposed to a private marketing offer.  Since they are an auto-loan company, they are curious about whether the model of car (e.g. Honda Accord, Toyota Corrola, Jeep Wrangler, etc.) being financed influences a customer's willingness to sign up for the credit card.  The logic is that people who like adventure travel might also like specific kinds of cars.  If it is true that cars are a reflection of their owners, then the company might expand its credit card offerings to take advantage of its car ownership data. 
+
+### An Initial Step
+
+Let's get started capturing the BankPass problem as a graphical model.  The first thought should be to make a simple model of how the real-world ends up generating your variable(s) of interest; and don't worry about how good the model is - just get something written.
+
+```{r cardModelSimple, echo = TRUE, out.width = "52%", fig.cap="Figure 1: Simple model.", fig.align="center"}
+library(causact)
+dag_create() %>%
+  dag_node("Get Card", "Y") %>%
+  dag_node("Adventurous Personality", "X",
+           child = "Y") %>%
+  dag_render(shortLabel = TRUE)
+
+```
+
+Figure 1 is a visual depiction of what mathematicians and computer scientists call a _graph_.  To them, a _graph_ is a set of related objects where the objects are called _nodes_ and any pair of related nodes are known as an _edge_.  For our purposes, a _node_ is a visual depiction of a random variable (i.e. an oval) and the presence of a visible _edge_ indicates a relationship between the connected random variables. 
+
+### Ensuring Nodes Are RV's {#analysis-of-nodes}
+
+Vetting the model (Figure 1) reveals some shortcomings and our first vetting criteria is to ensure the two nodes can be modelled as random variables with each node representing a mapping of real-world outcomes to a numerical values.  
+
+For the "Get Card" node, let's label it $X$, the mapping of outcomes to values is the conventional mapping for a Bernoulli-distributed random variable:
+
+$$
+X \equiv 
+\begin{cases}
+  0, & \textrm{if customer does not get the card }\\
+  1, & \textrm{if customer gets the TRC card }
+\end{cases}
+$$
+
[…]
