+++ DESCRIPTION
@@ -43,7 +43,10 @@ Imports:
     lifecycle
 RoxygenNote: 7.1.2
 Suggests: 
+    knitr,
     covr,
-    testthat (>= 3.0.0)
+    testthat (>= 3.0.0),
+    rmarkdown
 Config/testthat/edition: 3
 Roxygen: list(markdown = TRUE)
+VignetteBuilder: knitr
