+++ tests/testthat/test-rbern.R
@@ -1,0 +1,3 @@
+test_that("bernoulli rvs can be generated", {
+  expect_equal(sum(rbern(10,1)), 10)
+})
