### 2022-06-07 02c2143
attempt to fix github build issues with vignette
