+++ .github/workflows/test-coverage.yaml
@@ -1,29 +1,74 @@
-# Workflow derived from https://github.com/r-lib/actions/tree/master/examples
-# Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
 on:
   push:
-    branches: [main, master]
+    branches:
+      - main
+      - master
   pull_request:
-    branches: [main, master]
+    branches:
+      - main
+      - master
 
 name: test-coverage
 
 jobs:
   test-coverage:
-    runs-on: ubuntu-latest
+    runs-on: macOS-latest
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
-
     steps:
       - uses: actions/checkout@v2
 
       - uses: r-lib/actions/setup-r@v1
-        with:
-          use-public-rspm: true
 
-      - uses: r-lib/actions/setup-r-dependencies@v1
+      - uses: r-lib/actions/setup-pandoc@v1
+
+      - name: Query dependencies
+        run: |
+          install.packages('remotes')
+          saveRDS(remotes::dev_package_deps(dependencies = TRUE), ".github/depends.Rds", version = 2)
+          writeLines(sprintf("R-%i.%i", getRversion()$major, getRversion()$minor), ".github/R-version")
+        shell: Rscript {0}
+
+      - name: Restore R package cache
+        uses: actions/cache@v2
         with:
-          extra-packages: covr
+          path: ${{ env.R_LIBS_USER }}
+          key: ${{ runner.os }}-${{ hashFiles('.github/R-version') }}-1-${{ hashFiles('.github/depends.Rds') }}
+          restore-keys: ${{ runner.os }}-${{ hashFiles('.github/R-version') }}-1-
+
+      - name: Install dependencies
+        run: |
+          install.packages(c("remotes"))
+          remotes::install_deps(dependencies = TRUE)
+          remotes::install_cran("covr")
+        shell: Rscript {0}
+
+      ###
+      - name: Install Miniconda
+        run: |
+          install.packages(c("remotes", "keras"))
+          reticulate::install_miniconda()
+        shell: Rscript {0}
+
+      - name: Set options for conda binary for macOS
+        if: runner.os == 'macOS'
+        run: |
+          echo "options(reticulate.conda_binary = reticulate:::miniconda_conda())" >> .Rprofile
+#  Perhaps here is where we can install / change the environment that we are
+# installing into? Can we call our own greta install functions here?
+      - name: Install TensorFlow
+        run: |
+          reticulate::conda_create(envname = "greta-env",python_version = "3.7")
+          reticulate::conda_install(envname = "greta-env", packages = c("numpy==1.16.4", "tensorflow-probability==0.7.0", "tensorflow==1.14.0"))
+        shell: Rscript {0}
+
+      - name: Python + TF details
+        run: |
+          Rscript -e 'tensorflow::tf_config()'
+          Rscript -e 'tensorflow::tf_version()'
+          Rscript -e 'reticulate::py_module_available("tensorflow_probability")'
+          Rscript -e 'reticulate::py_config()'
+      ###
 
       - name: Test coverage
         run: covr::codecov()
