+++ .github/workflows/R-CMD-check.yaml
@@ -26,34 +26,3 @@ jobs:
           extra-packages: rcmdcheck
 
       - uses: r-lib/actions/check-r-package@v1
-
-      - name: Fix Conda permissions on macOS
-        if: runner.os == 'macOS'
-        run: sudo chown -R $UID $CONDA
-
-      - name: Remotes install greta
-        run: |
-          install.packages('remotes')
-          remotes::install_github("greta-dev/greta")
-        shell: Rscript {0}
-
-      - name: Greta install deps
-        run: |
-          library(greta)
-          install_greta_deps(timeout = 50)
-        shell: Rscript {0}
-
-      - name: Check the installation happened properly?
-        run: |
-          greta:::have_greta_conda_env()
-          greta:::have_conda()
-          greta:::have_python()
-          greta:::have_tf()
-          greta:::have_tfp()
-        shell: Rscript {0}
-
-      - name: Greta run model
-        run: |
-          library(greta)
-          model(normal(0,1))
-        shell: Rscript {0}
