### 2022-06-07 8baa8af
update test-coverage to exclude Tensorflow dependency install
