### 2022-06-09 cf938cf
minor updates for CRAN submission
