+++ README.md
@@ -126,16 +126,16 @@ drawsDF  ### see top of data frame
 #> # A tibble: 4,000 x 4
 #>    theta_JpWrnglr theta_KiaForte theta_SbrOtbck theta_ToytCrll
 #>             <dbl>          <dbl>          <dbl>          <dbl>
-#>  1          0.819          0.229          0.652          0.213
-#>  2          0.819          0.229          0.652          0.213
-#>  3          0.819          0.229          0.652          0.213
-#>  4          0.841          0.306          0.657          0.199
-#>  5          0.841          0.306          0.657          0.199
-#>  6          0.839          0.229          0.582          0.192
-#>  7          0.856          0.243          0.638          0.206
-#>  8          0.856          0.243          0.638          0.206
-#>  9          0.847          0.238          0.579          0.197
-#> 10          0.868          0.192          0.657          0.197
+#>  1          0.811          0.242          0.605          0.199
+#>  2          0.818          0.288          0.622          0.200
+#>  3          0.819          0.175          0.613          0.171
+#>  4          0.817          0.228          0.617          0.213
+#>  5          0.819          0.166          0.601          0.176
+#>  6          0.819          0.265          0.614          0.226
+#>  7          0.819          0.205          0.613          0.209
+#>  8          0.822          0.204          0.582          0.244
+#>  9          0.817          0.196          0.597          0.202
+#> 10          0.816          0.172          0.610          0.193
 #> # ... with 3,990 more rows
 ```
 
