+++ README.md
@@ -126,16 +126,16 @@ drawsDF  ### see top of data frame
 #> # A tibble: 4,000 x 4
 #>    theta_JpWrnglr theta_KiaForte theta_SbrOtbck theta_ToytCrll
 #>             <dbl>          <dbl>          <dbl>          <dbl>
-#>  1          0.811          0.242          0.605          0.199
-#>  2          0.818          0.288          0.622          0.200
-#>  3          0.819          0.175          0.613          0.171
-#>  4          0.817          0.228          0.617          0.213
-#>  5          0.819          0.166          0.601          0.176
-#>  6          0.819          0.265          0.614          0.226
-#>  7          0.819          0.205          0.613          0.209
-#>  8          0.822          0.204          0.582          0.244
-#>  9          0.817          0.196          0.597          0.202
-#> 10          0.816          0.172          0.610          0.193
+#>  1          0.823          0.205          0.678          0.190
+#>  2          0.855          0.220          0.637          0.200
+#>  3          0.826          0.224          0.582          0.223
+#>  4          0.831          0.212          0.629          0.173
+#>  5          0.878          0.236          0.619          0.210
+#>  6          0.801          0.277          0.610          0.199
+#>  7          0.878          0.240          0.633          0.208
+#>  8          0.878          0.240          0.633          0.208
+#>  9          0.874          0.275          0.616          0.214
+#> 10          0.870          0.301          0.589          0.224
 #> # ... with 3,990 more rows
 ```
 
