+++ vignettes/narrative-to-insight-with-causact.Rmd
@@ -19,7 +19,7 @@ library(causact)
 
 ## Motivating Business Narrative
 
-We will use the below example fictional narrative along with some simulated data to motivate how to use use `causact` for iterating from narrative to computational insight.  A more detailed version of this vignette can be found in Chapters 14-17 of "A Business Analyst's Introduction to Business Analytics" available at http://www.causact.com.
+We will use the below example fictional narrative along with some simulated data to motivate how to use use `causact` for iterating from narrative to computational insight.  A more detailed version of this vignette can be found in Chapters 14-17 of "A Business Analyst's Introduction to Business Analytics" available at https://www.causact.com.
 
 > BankPass is an auto financing company that is launching a new credit card, the Travel Rewards Card (TRC).  The card is designed for people who love to take adventure vacations like ziplining, kayaking, scuba diving, and the like.  To date, BankPass has been randomly selecting a small number of customers to target with a private offer to sign up for the card.  They would like to assess the probability that an individual will get a TRC card if exposed to a private marketing offer.  Since they are an auto-loan company, they are curious about whether the model of car (e.g. Honda Accord, Toyota Corrola, Jeep Wrangler, etc.) being financed influences a customer's willingness to sign up for the credit card.  The logic is that people who like adventure travel might also like specific kinds of cars.  If it is true that cars are a reflection of their owners, then the company might expand its credit card offerings to take advantage of its car ownership data.
 
@@ -37,7 +37,7 @@ dag_create() %>%
   dag_render(shortLabel = TRUE)
 ```
 
-<img src="figure/cardModelSimple-1.png" title="Figure 1: Simple model." alt="Figure 1: Simple model." width="52%" style="display: block; margin: auto;" />
+<img src="cardModelSimple-1.png" title="Figure 1: Simple model." alt="Figure 1: Simple model." width="52%" style="display: block; margin: auto;" />
 
 The above output is a visual depiction of what mathematicians and computer scientists call a _graph_.  To them, a _graph_ is a set of related objects where the objects are called _nodes_ and any pair of related nodes are known as an _edge_.  For our purposes, a _node_ is a visual depiction of a random variable (i.e. an oval) and the presence of a visible _edge_ indicates a relationship between the connected random variables.
 
@@ -69,7 +69,7 @@ dag_create() %>%
   dag_render(shortLabel = FALSE)
 ```
 
-<img src="figure/cardModelSimpleRevised-1.png" title="Figure 2: Revised simple model." alt="Figure 2: Revised simple model." width="52%" style="display: block; margin: auto;" />
+<img src="cardModelSimpleRevised-1.png" title="Figure 2: Revised simple model." alt="Figure 2: Revised simple model." width="52%" style="display: block; margin: auto;" />
 
 As shown above, the difficult to measure personality node gets replaced by the car model node.  To verify we can measure _Car Model_ as a random variable, let's be explicit about the mapping of "Car Model" outcomes to real numbers.  For example:
 
@@ -131,7 +131,7 @@ graph = dag_create() %>%
 graph %>% dag_render()
 ```
 
-<img src="figure/theta44-1.png" title="Figure 3: A generative DAG for the credit card example." alt="Figure 3: A generative DAG for the credit card example." width="52%" style="display: block; margin: auto;" />
+<img src="theta44-1.png" title="Figure 3: A generative DAG for the credit card example." alt="Figure 3: A generative DAG for the credit card example." width="52%" style="display: block; margin: auto;" />
 
[…]
