+++ cran-comments.md
@@ -1,13 +1,13 @@
-This is a patch release to accommodate breaking changes coming from `tidyverse::tidyr`.
+This is a patch release to accommodate `greta::variable` as a valid distribution for flat priors, to include a vignette as part of the package, and to automate some unit tests.
 
 ## Test environments
-* local R installation, R 4.0.0
-* ubuntu 16.04 (on travis-ci), R 4.0.0
+* local R installation, R 4.1.1
+* ubuntu 16.04 (on travis-ci), R 4.1.1
 * win-builder (devel)
 
 ## R CMD check results
 
-0 errors | 0 warnings | 1 note
+0 errors | 0 warnings | 0 note
 
 ## rhub::check_for_cran() 
 
