### 2022-06-09 bc12b13
Increment version number to 0.4.2
