### 2022-06-10 #4415 @zhiiiyang
@terrytangyuan, I just found out that I can't use this package since it depends on `greta` failing to run on an M1 Mac。:(. This has been confirmed by the maintainer. Could you remove me from the reviewer list? Sorry about the inconveniences.
https://github.com/greta-dev/greta/issues/526#issuecomment-1151903419
