### 2022-06-14 ff7e210
fix broken link in README.md for CRAN submission
