+++ cran-comments.md
@@ -1,5 +1,8 @@
 This is a patch release to accommodate `greta::variable` as a valid distribution for flat priors, to include a vignette as part of the package, and to automate some unit tests.
 
+## Resubmission
+This is a resubmission to CRAN that fixes a broken link in the README.md.
+
 ## Test environments
 * local R installation, R 4.1.1
 * ubuntu 16.04 (on travis-ci), R 4.1.1
