+++ NEWS.md
@@ -1,3 +1,5 @@
+# causact (development version)
+
 # causact 0.4.2
 * Added `vignette("narrative-to-insight-with-causact") to introduce the package's encouraged user workflow.
 * Allow `greta::variable()` to be used for flat priors.
