### 2022-06-14 flyaflya/causact#43 @flyaflya
@zhiiiyang Thanks so much for agreeing to review.  I am sorry you faced the M1 chip issue.  I tried to alert reviewers to that issue (see here: https://github.com/openjournals/joss-reviews/issues/4358#issuecomment-1135084232), but this message got lost when the new review got started.

So, even though you are removing yourself as a reviewer due to the M1 issue, I do appreciate your comments above and have fixed them all.

Community guidelines:  https://github.com/flyaflya/causact/blob/master/CODE_OF_CONDUCT.md
Issue/Bug Support Instructions are now in the README.md:
![image](https://user-images.githubusercontent.com/17184441/173626915-880db53f-1138-41aa-a4d3-6aa18b60c7c7.png)
Contributor guide: https://github.com/flyaflya/causact/blob/master/.github/contributing.md

Unit Tests: https://github.com/flyaflya/causact/tree/master/tests/testthat

Vignette: https://cran.r-project.org/web/packages/causact/vignettes/narrative-to-insight-with-causact.html

Thanks again for the help and agreeing to review!  It is much aprreciated.
