### 2022-06-14 #4415 @flyaflya
@terrytangyuan Even though @zhiiiyang is no longer an official reviewer, her comments/suggestions were much appreciated and there is a new CRAN release of `causact` as of today to support adding a vignette.  You can see more details about my response to @zhiiiyang 's review here: https://github.com/flyaflya/causact/issues/43#issuecomment-1155419729
