### 2022-07-19 #4415 @flyaflya
Hi all - just wanted to gently check on the progress of this review.   @ChristopherLucas does https://github.com/openjournals/joss-reviews/issues/4415#issuecomment-1147543764 help ease your concerns about the presented model in the paper?  @terrytangyuan is there any additional outstanding comments/issues that I can help address?

Thanks to all for the help!
