### 2022-07-19 #4415 @arfon
:wave: folks. As @terrytangyuan is stepping down from the JOSS editorial team (thanks for all of your help over the years Yuan!) I'm going to be taking over this review.

Based on my reading of the review thread thus far, it looks like there has been some feedback from @flyaflya in response to @ChristopherLucas' feedback. @ChristopherLucas – could you take another look and see if you're able to complete your review at this stage?

Many thanks!
