### 2022-07-26 #4415 @ChristopherLucas
> @ChristopherLucas A nearly equivalent representation of the eight-schools problem can be shown like this: ![image](https://user-images.githubusercontent.com/17184441/172177380-73c916d5-2918-4e3a-a484-866e249f6824.png) (`variable()` represents a flat prior). This picture above exactly matches the description in Gelman et al. (2014) p. 119-120 (available here: http://www.stat.columbia.edu/~gelman/book/BDA3.pdf). If this representation seems more helpful, I can use it in place of the one that is there already?
>
> I am with you that typically the "the treatment effect (y)" should not be "dependent on my estimates of parameters (i.e. sigma) of the sampling distribution of the estimator." The only reason the eight-schools problem looks like this is due to this key assumption articulated in Gelman: "the sampling variances [i.e. the std error of effect] are known, for all practical purposes." Thus, the measured treatment effect is dependent on the assumed-to-be-known parameters of the sampling distribution of the estimator.
>
> I could use a different dataset/model to showcase `causact`, but this dataset was used because it commonly appears in the computational Bayesian inference literature and the graphical illustration of the model helps to initiate conversations about assumptions like the conversation we are having here. Examples of eight schools used as demo model:
>
>     * Sturtz, S., Ligges, U., & Gelman, A. (2005). R2WinBUGS: a package for running WinBUGS from R. Journal of Statistical software, 12, 1-16 (https://www.jstatsoft.org/index.php/jss/article/view/v012i03/33).
>
>     * `rstan` Vignette:  https://cran.r-project.org/web/packages/rstan/vignettes/rstan.html
>
>
> PS: The justification for the simplifiying asumption that the estimated sigma from the data is assumed to be equivalent to the "true" sigma of the sampling distribution is because the treatment effect (y) is actually an observed mean treatment effect over many (>30) individual student performances at each school. Presumably, if you run the experiment with a different set of students, the sampling distribution of their mean treatment effect y∼Normal(theta,sigma).

This all makes sense to me. I still worry that researchers regularly working with DAGs will find it confusing to mix the causal structure with assumptions about the estimator for estimands denoted by that structure. I think the authors know precisely what they mean by the example, and that they mean something sensible, but I do think it will be unclear to many researchers who regularly work with DAGs but not with Bayesian models. I'd probably use a simpler example and perhaps invoke the 8-schools problem elsewhere in the docs, where there is more space to clarify what's going on. E.G., if I were to set up an example of the eight-schools problem, I'd write the estimand in potential outcomes notation, then denote a DAG (or several DAGs) under which a particular estimator returns an estimate of that estimand with some characterizable properties wrt bias/variance. That's probably a bit outside the bounds of a software submission, so I don't think it necessary here, but I do worry that as written many interested readers will be unnecessarily confused.
