### 2022-07-26 bb4f171
Change example from eight schools to chimp example.
