+++ paper/paper.md
@@ -50,50 +50,60 @@ To unify inference-problem narratives, the statistical models representing those
 
 Generative DAGs pursue two simultaneous goals.  One goal is to capture the narrative by building a conceptual understanding of the data generating process that lends itself to statistical modelling.  And the second goal is to gather all the mathematical elements needed for specifying a complete Bayesian model of the data generating process.  Both of these goals will be satisfied by iteratively assessing the narrative and the narrative's translation into rigorous mathematics using `causact` functions.
 
-Capturing the narrative in code uses some core `causact` functions like `dag_create()`, `dag_node()`, `dag_edge()`, and `dag_plate()` with the chaining operator `%>%` used to build a DAG from the individual elements.  `dag_render()` or `dag_greta()` are then used to visualize the DAG or run inference on the DAG, respectively.  The simplicity with which generative DAGs are constructed belies the complexity of models which can be supported.  For example, multi-level or hierarchical models are easily constructed as shown here in code for constructing and visualizing an oft-cited Bayesian example known as eight schools [@JSSv012i03] whose data is included in `causact` (`causact::schoolsDF`).  The example is a study of coaching effects on test scores where students from eight schools were put into coached and uncoached groups.
+Capturing the narrative in code uses some core `causact` functions like `dag_create()`, `dag_node()`, `dag_edge()`, and `dag_plate()` with the chaining operator `%>%` used to build a DAG from the individual elements.  `dag_render()` or `dag_greta()` are then used to visualize the DAG or run inference on the DAG, respectively.  The simplicity with which generative DAGs are constructed belies the complexity of models which can be supported.  For example, multi-level or hierarchical models are easily constructed as shown here in code for constructing and visualizing an experiment about chimpanzees [@mcelreath2020statistical] whose data is included in `causact` (`causact::chimpanzeesDF`).  Chimpanzees are given a choice to pull one of two levers - right or left.  Depending on the lever pulled, the chimpanzee is either prosocial, pulling the lever which feeds both himself and a partner, or not prosocial where only the lever-puller receives food.  
 
 ```r
-graph = dag_create() %>%
-    dag_node("Treatment Effect","y",
-             rhs = normal(theta, sigma),
-             data = causact::schoolsDF$y) %>%
-    dag_node("Std Error of Effect Estimates","sigma",
-             data = causact::schoolsDF$sigma,
-             child = "y") %>%
-    dag_node("Exp. Treatment Effect","theta",
-             child = "y",
-             rhs = avgEffect + schoolEffect) %>%
-    dag_node("Population Treatment Effect","avgEffect",
-             child = "theta",
-             rhs = normal(0,30)) %>%
-    dag_node("School Level Effects","schoolEffect",
-             rhs = normal(0,30),
-             child = "theta") %>%
-    dag_plate("Observation","i",nodeLabels = c("sigma","y","theta")) %>%
-    dag_plate("School Name","school",
-              nodeLabels = "schoolEffect",
-              data = causact::schoolsDF$schoolName,
-              addDataNode = TRUE)
+## get only the experiments with partner present
+chimpDF = causact::chimpanzeesDF %>% filter(condition == 1)
+## create model
+graph = dag_create() %>% 
[…]
