### 2022-07-26 afcbf96
change figure widths
