### 2022-07-27 #4415 @flyaflya
Thanks [@ChristopherLucas](https://github.com/ChristopherLucas) for the encouragement. I switched the example model to use another well known dataset without the weird modelling of standard errors. This time I am using a medium complexity version of the chimpanzee example from McElreath's Statistical Rethinking. Experiment discussed here: https://youtu.be/n2aJYtuGu54?t=513

The updated paper is an artifact from the last action and can be downloaded here:
https://github.com/flyaflya/causact/suites/7535235027/artifacts/310574849
