### 2022-07-28 #4415 @flyaflya
Fixed a few minor typos and the most recent manuscript can be found as the artifact here: https://github.com/flyaflya/causact/actions/runs/2743355474

@arfon, thanks for taking over as editor.  I believe the review process is nearing its conclusion (checklists complete and all issues have been addressed), so please let me know what I else I can do to help shepherd things along.  Thanks again!
