### 2022-08-01 e8eb4da
Fix Figure 3 miscaptioning for JOSS release.
