### 2022-08-01 ccf8c53
Figure 3 tweaks for publication.
