+++ paper/joss.bib
@@ -41,7 +41,7 @@
 	year         = 2020,
 	publisher    = {Independently Published},
 	isbn         = 9798667128175,
-	url          = {https://www.causact.com/}
+	note          = {\url{https://www.causact.com/}}
 }
 @article{gabry2019visualization,
 	title        = {Visualization in {Bayesian} workflow},
